+++ CHANGELOG.md
@@ -11,19 +11,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update to Miniconda version to 24.5.0-0
-- Make a symlink of `f2py` to `f2py3` to fix an issue with the wrong `f2py` being found sometimes on discover
-
 ### Added
-
-- Explicit Conda Packages
-  - uxarray
-- Explicit Pip Packages
-  - basemap
 
 ### Removed
 
 ### Deprecated
+
+## [24.7.0] - 2024-09-10
+
+### Changed
+
+- Migrated to use Miniforge instead of Miniconda
+- "Update" to Miniforge version to 24.7.1-0
+  - Note: Miniforge is often behind Miniconda in version numbers, so this is not a downgrade, but rather
+    the latest version of Miniforge
+- Make a symlink of `f2py` to `f2py3` to fix an issue with the wrong `f2py` being found sometimes on discover
+- Update both `.condarc` and `.mambarc` to use the `conda-forge` channel as the default channel
+- Add a test at the end to ensure no `defaults` channel packages are installed
+- Add `-ffnet-hack` for a hack to install `ffnet` on bucy (nothing else seems to need it)
+  - This flag uses a fork of ffnet with hardcoded references to gfortran
+
+### Added
+
+- Explicit Mamba Packages
+  - uxarray
+- Explicit Pip Packages
+  - basemap
+  - redis (moved from Mamba to Pip)
+- Added basemap example
+
+### Removed
+
+- Removed `--conda` option
+- Explicit Mamba Packages
+  - redis (moved from Mamba to Pip)
+  - redis-py (seems not to exist in the conda-forge channel or in PyPI)
+  - sphinxcontrib (seems not to exist in the conda-forge channel or in PyPI)
+  - get_terminal_size (built-in to Python 3.3+ in the `shutil` module)
 
 ## [24.4.0] - 2024-05-30